feat: uses GeckoView now & fixed ad issue (removed)
